AdbHelper: Move read/write code to the AdbSocket class.
